requirements.txt
Django>=4.1.1,<4.2
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
requests>=2.28.1,<2.29
